Extract core socket tests
